+++ src/main/java/no/vegvesen/nvdbapi/client/gson/PlacementParser.java
@@ -48,11 +48,7 @@ public final class PlacementParser {
     }
 
     private static RoadPlacement parseRoadPlacement(JsonObject obj) {
-        RoadSysRef roadRef = RoadSysRefParser.parse(obj.getAsJsonObject("vegsystemreferanse"));
-        RefLinkPosition refLink = ShortRefLinkParser.parseShortRefLink(obj.getAsJsonObject("veglenkesekvens"));
-        Geometry point = GeometryParser.parse(obj.getAsJsonObject("geometri"));
-
-        return new RoadPlacement(roadRef, refLink, point);
+        return RoadPlacementParser.parseRoadPlacement(obj);
     }
 
     public static Position.Result parsePosition(JsonObject obj) {